feat: Set CRF value based on video height in Converter class
The commit updates the `Converter` class in `converter.py` to set the CRF (Constant Rate Factor) value based on the height of the video in pixels. The default CRF value is 28, but for videos with a height of 500 pixels or less, the CRF value is set to 23. This change improves the video compression quality for smaller videos.